// admin/src/api/generated/models/CreateUserRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * deploy-go-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface CreateUserRequest
 */
export interface CreateUserRequest {
    /**
     *
     * @type {string}
     * @memberof CreateUserRequest
     */
    displayName?: string | null;
    /**
     *
     * @type {string}
     * @memberof CreateUserRequest
     */
    email?: string | null;
    /**
     *
     * @type {string}
     * @memberof CreateUserRequest
     */
    password: string;
    /**
     *
     * @type {string}
     * @memberof CreateUserRequest
     */
    username: string;
}

/**
 * Check if a given object implements the CreateUserRequest interface.
 */
export function instanceOfCreateUserRequest(value: object): value is CreateUserRequest {
    if (!('password' in value) || value['password'] === undefined) return false;
    if (!('username' in value) || value['username'] === undefined) return false;
    return true;
}

export function CreateUserRequestFromJSON(json: any): CreateUserRequest {
    return CreateUserRequestFromJSONTyped(json, false);
}

export function CreateUserRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateUserRequest {
    if (json == null) {
        return json;
    }
    return {

        'displayName': json['display_name'] === undefined ? undefined : json['display_name'] === null ? null : json['display_name'],
        'email': json['email'] === undefined ? undefined : json['email'] === null ? null : json['email'],
        'password': json['password'],
        'username': json['username'],
    };
}

export function CreateUserRequestToJSON(json: any): CreateUserRequest {
    return CreateUserRequestToJSONTyped(json, false);
}

export function CreateUserRequestToJSONTyped(value?: CreateUserRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'display_name': value['displayName'],
        'email': value['email'],
        'password': value['password'],
        'username': value['username'],
    };
}